Code update

+++ _common/s3helpers.py
@@ -20,6 +20,7 @@ class S3Handler:
         client.download_file(bucket, s3_path, filename)
         return filename
 
+
     @staticmethod
     def download_folder(bucket: str, s3_path: str, folder: Path, report=None):
         if os.path.exists(folder.__str__()):
@@ -33,12 +34,17 @@ class S3Handler:
         if report == 'tqdm':
             keys = keys.feed(fluq.with_progress_bar())
 
+        client = boto3.client('s3')
+
         for key in keys:
             proper_key = key[len(s3_path):]
             if proper_key.startswith('/'):
                 proper_key=proper_key[1:]
             filename = folder.joinpath(proper_key)
-            S3Handler.download_file(bucket, key, filename)
+            os.makedirs(str(filename.parent), exist_ok=True)
+            filename = filename.__str__()
+            client.download_file(bucket, key, filename)
+
 
     @staticmethod
     def upload_file(bucket_name: str, s3_path: str, filename: Union[Path,str]):
@@ -55,7 +61,6 @@ class S3Handler:
 
     @staticmethod
     def upload_folder(bucket_name: str, s3_path: str, folder: Path):
-
         aws_access_key_id = os.environ.get('AWS_ACCESS_KEY_ID', None)
         aws_secret_access = os.environ.get('AWS_SECRET_ACCESS_KEY', None)
         if aws_access_key_id is not None and aws_secret_access is not None:
@@ -68,10 +73,12 @@ class S3Handler:
         bucket.objects.filter(Prefix=s3_path).delete()
 
         client = boto3.client('s3', **kwargs)
-        for file_path in Query.folder(folder):
-            file_path_str = file_path.__str__()
-            joint_path = os.path.join(s3_path, file_path.name)
-            client.upload_file(file_path_str, bucket_name, joint_path)
+        for file_path in Query.folder(folder,'**/*'):
+            if file_path.is_file():
+                file_path_str = file_path.__str__()
+                relative_path= file_path.relative_to(folder)
+                joint_path = os.path.join(s3_path, relative_path)
+                client.upload_file(file_path_str, bucket_name, joint_path)
 
 
     @staticmethod
@@ -90,3 +97,27 @@ class S3Handler:
             print(path)
         else:
             bucket.objects.filter(Prefix=path).delete()
+
+    @staticmethod
+    def get_latest_modified_file(bucket_name: str, path: str):
+        aws_access_key_id = os.environ.get('AWS_ACCESS_KEY_ID', None)
+        aws_secret_access = os.environ.get('AWS_SECRET_ACCESS_KEY', None)
+        if aws_access_key_id is not None and aws_secret_access is not None:
+            kwargs = dict(aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access)
+        else:
+            kwargs = {}
+
+        s3 = boto3.client('s3', **kwargs)
+
+        response = s3.list_objects_v2(
+            Bucket=bucket_name,
+            Prefix=path,
+        )
+        if response['ResponseMetadata']['HTTPStatusCode'] == 200 and len(response['Contents']) > 0:
+            max_modified = response['Contents'][0]['LastModified']
+            path = response['Contents'][0]['Key']
+            for object in response['Contents']:
+                if object['LastModified'] > max_modified:
+                    path = object['Key']
+            return path
+        return None
